Move typed interface to own module
